[MODINVOICE-132] - Batch voucher generation and persistence Refactoring
