+++ src/main/java/org/folio/invoices/utils/HelperUtils.java
@@ -22,6 +22,7 @@ import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -32,6 +33,7 @@ import java.util.concurrent.CompletionException;
 import java.util.function.Predicate;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.IntStream;
 
 import javax.money.CurrencyUnit;
 import javax.money.Monetary;
@@ -467,5 +469,19 @@ public class HelperUtils {
   public static Money getAdjustmentFundDistributionAmount(FundDistribution fundDistribution, Adjustment adjustment, Invoice invoice) {
     MonetaryAmount adjustmentTotal = calculateAdjustment(adjustment, Money.of(invoice.getSubTotal(), invoice.getCurrency()));
     return getFundDistributionAmount(fundDistribution, adjustmentTotal.getNumber().doubleValue(), invoice.getCurrency());
+  }
+
+  public static <T> Map<Integer, List<T>> buildIdsChunks(List<T> source, int maxListRecords) {
+    int size = source.size();
+    if (size <= 0)
+      return Collections.emptyMap();
+    int fullChunks = (size - 1) / maxListRecords;
+    HashMap<Integer, List<T>> idChunkMap = new HashMap<>();
+    IntStream.range(0, fullChunks + 1)
+      .forEach(n -> {
+        List<T> subList = source.subList(n * maxListRecords, n == fullChunks ? size : (n + 1) * maxListRecords);
+        idChunkMap.put(n, subList);
+      });
+    return idChunkMap;
   }
 }
