+++ src/main/java/org/folio/rest/impl/BatchVoucherExportsHelper.java
@@ -75,16 +75,21 @@ public class BatchVoucherExportsHelper extends AbstractHelper {
    * @return completable future with {@link BatchVoucherExport} on success or an exception if processing fails
    */
   public CompletableFuture<BatchVoucherExport> createBatchVoucherExports(BatchVoucherExport batchVoucherExport) {
-    return createRecordInStorage(JsonObject.mapFrom(batchVoucherExport), resourcesPath(BATCH_VOUCHER_EXPORTS_STORAGE))
-                  .thenApply(batchVoucherExport::withId)
-                  .thenApply(this::persistBatchVoucher);
+    CompletableFuture<BatchVoucherExport> future = new CompletableFuture<>();
+    createRecordInStorage(JsonObject.mapFrom(batchVoucherExport), resourcesPath(BATCH_VOUCHER_EXPORTS_STORAGE))
+                  .thenApply(batchVoucherExportId -> {
+                    BatchVoucherExport batchVoucherExportWitId = batchVoucherExport.withId(batchVoucherExportId);
+                    future.complete(batchVoucherExport.withId(batchVoucherExportId));
+                    return batchVoucherExportWitId;
+                  })
+                  .thenAccept(this::persistBatchVoucher);
+    return future;
   }
 
-  private BatchVoucherExport persistBatchVoucher(BatchVoucherExport batchVoucherExport) {
+  private void persistBatchVoucher(BatchVoucherExport batchVoucherExport) {
     VertxCompletableFuture.runAsync(ctx,
       () -> sendEvent(MessageAddress.BATCH_VOUCHER_PERSIST_TOPIC
               , new JsonObject().put(BATCH_VOUCHER_EXPORT, JsonObject.mapFrom(batchVoucherExport))));
-    return batchVoucherExport;
   }
 
   /**
